+++ rasa/core/tracker_store.py
@@ -166,16 +166,16 @@ class RedisTrackerStore(TrackerStore):
         host="localhost",
         port=6379,
         db=0,
-        password=None,
-        event_broker=None,
-        record_exp=None,
-        ssl=False,
+        password: Optional[Text] = None,
+        event_broker: Optional[EventChannel] = None,
+        record_exp: Optional[float] = None,
+        use_ssl: bool =False,
     ):
 
         import redis
 
         self.red = redis.StrictRedis(
-            host=host, port=port, db=db, password=password, ssl=ssl
+            host=host, port=port, db=db, password=password, ssl=use_ssl
         )
         self.record_exp = record_exp
         super(RedisTrackerStore, self).__init__(domain, event_broker)